requirements.txt
opencv-python==3.4.4.19
numpy==1.15.4
face_recognition==1.2.3
dlib==19.16.0
Django==2.1.4
djangorestframework==3.9.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
django-extensions==2.1.4
requests==2.21.0
scikit-image==0.14.1
face-alignment==1.0.0
celery==4.2.1
amqp==2.3.2
flower==0.9.2
kombu==4.2.2.post1
uWSGI==2.0.17.1
django-timezone-field==3.0